+++ src/main/java/me/Tamaninja/test/entity/PalletContainer.java
@@ -11,16 +11,16 @@ public class PalletContainer implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "palletContainer_id", nullable = false, unique = true)
-    private Long id;
+    private Integer id;
     @Column(name = "palletContainer_name", nullable = false, unique = true)
     private String name;
     @Column(name = "palletContainer_weight", nullable = false)
     private double weight;
 
     @Column(name = "palletContainer_amount", nullable = false)
-    private int defaultAmount;
+    private Short defaultAmount;
 
-    public PalletContainer(String name, double weight, int defaultAmount) {
+    public PalletContainer(String name, double weight, Short defaultAmount) {
         this.name = name;
         this.weight = weight;
         this.defaultAmount = defaultAmount;
@@ -34,7 +34,7 @@ public class PalletContainer implements Serializable {
     public double getWeight() {
         return (weight);
     }
-    public int getDefaultAmount() {
+    public Short getDefaultAmount() {
         return (defaultAmount);
     }
