+++ src/main/java/me/Tamaninja/test/repository/PalletContainerRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PalletContainerRepo extends JpaRepository<PalletContainer,Long> {
+public interface PalletContainerRepo extends JpaRepository<PalletContainer,Integer> {
 
 }
