+++ src/main/java/me/Tamaninja/test/repository/PalletTypeRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PalletTypeRepo extends JpaRepository<PalletType,Long> {
+public interface PalletTypeRepo extends JpaRepository<PalletType,Integer> {
 
 }
