+++ src/main/java/me/Tamaninja/test/repository/TruckRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TruckRepo extends JpaRepository<Truck, Long> {
+public interface TruckRepo extends JpaRepository<Truck, Integer> {
 
 }
